Fin TP8 - Modificaciones de Login

+++ Controllers/ProductosController.cs
@@ -21,12 +21,20 @@ public class ProductosController : Controller
     [HttpGet]
     public IActionResult CrearProducto()
     {
+        if (HttpContext.Session.GetString("rol") != "admin")
+        {
+            return RedirectToAction("Index", "Home");
+        }
         return View(new Productos());
     }
 
     [HttpPost]
     public IActionResult CrearProducto(Productos productoNuevo)
     {
+        if (HttpContext.Session.GetString("rol") != "admin")
+        {
+            return RedirectToAction("Index", "Home");
+        }
         productosRep.CreateProducto(productoNuevo);
         return RedirectToAction("Index");
     }
@@ -34,12 +42,20 @@ public class ProductosController : Controller
     [HttpGet]
     public IActionResult ModificarProducto(int idBuscado)
     {
+        if (HttpContext.Session.GetString("rol") != "admin")
+        {
+            return RedirectToAction("Index", "Home");
+        }
         return View(productosRep.GetAllProductos().FirstOrDefault(prod => prod.IdProducto == idBuscado));
     }
 
     [HttpPost]
     public IActionResult ModificarProducto(Productos productoBuscado)
     {
+        if (HttpContext.Session.GetString("rol") != "admin")
+        {
+            return RedirectToAction("Index", "Home");
+        }
         productosRep.UpdateProducto(productoBuscado.IdProducto, productoBuscado);
         return RedirectToAction("Index");
     }
@@ -47,12 +63,20 @@ public class ProductosController : Controller
     [HttpGet]
     public IActionResult EliminarProducto(int idProducto)
     {
+        if (HttpContext.Session.GetString("rol") != "admin")
+        {
+            return RedirectToAction("Index", "Home");
+        }
         return View(productosRep.GetAllProductos().FirstOrDefault(prod => prod.IdProducto == idProducto));
     }
 
     [HttpPost]
     public IActionResult EliminarProducto(Productos productoBuscado)
     {
+        if (HttpContext.Session.GetString("rol") != "admin")
+        {
+            return RedirectToAction("Index", "Home");
+        }
         productosRep.DeleteProducto(productoBuscado.IdProducto);
         return RedirectToAction("Index");
     }
